src/requirements.txt
SQLAlchemy
alembic
psycopg2-binary
python-dotenv
asyncpg
greenlet
fastapi
uvicorn
mnemonic
ecdsa
cryptography
PyNaCl
bech32
base64